Cinic-10_V1_FlyAI/requirements.txt
flyai
numpy
Pillow
torch
albumentations
